+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.3] - 2019-11-26
 ### Removed
 - Fields enableInStorePickup & enableShipFromStore from location settings
 - Unnecessary constructors for attribute and customer feeds
@@ -172,7 +174,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - createCategory event & direct call
 - deleteCategory event & direct call
 
-[Unreleased]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.2...HEAD
+[Unreleased]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.3...HEAD
+[1.1.3]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/shopgate/connect-integration-sdk/compare/v1.0.0...v1.1.0